REFACT :: 패키지 조정, 주석 추가
