+++ src/main/kotlin/com/info/maeumgagym/infrastructure/filter/config/SecurityFilterChainConfig.kt
@@ -3,8 +3,8 @@ package com.info.maeumgagym.infrastructure.filter.config
 import com.info.maeumgagym.infrastructure.error.filter.ErrorLogResponseFilter
 import com.info.maeumgagym.infrastructure.error.filter.ExceptionConvertFilter
 import com.info.maeumgagym.security.authentication.provider.UserModelAuthenticationProvider
-import com.info.maeumgagym.security.jwt.MaeumgagymTokenAuthenticateFilter
-import com.info.maeumgagym.security.jwt.MaeumgagymTokenResolver
+import com.info.maeumgagym.security.mgtoken.filter.MaeumgagymTokenAuthenticateFilter
+import com.info.maeumgagym.security.mgtoken.MaeumgagymTokenResolver
 import org.springframework.security.config.annotation.SecurityConfigurerAdapter
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
 import org.springframework.security.web.DefaultSecurityFilterChain
